+++ lib/element/element_param.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+///
+/// Author Masahide Mori
+///
+/// First edition creation date 2022-01-30 20:19:25
+///
+class ScrollElementParams {
+  ScrollController? ctrl;
+}